Add files via upload
Deploying ELK Stack on Docker Container..
DESCRIPTION

Project objective: 

You have to deploy ELK Stack on a Docker container to implement continuous monitoring.

 

Background of the problem statement:  

Your manager has asked to create an elegant user interface for data analysis and data visualization as you have worked on ELK stack previously and have the idea of how it works. This will help the DevOps team to monitor and analyze the application behavior.

 

You must use the following: 

Docker: To create a container for installing ELK Stack

ELK Stack: To implement continuous monitoring

Git: To connect and push files from the local system to GitHub 

GitHub: To store the Angular application

 

Following requirements should be met: 

A few of the source code should be tracked on GitHub repositories. You need to document the tracked files that are ignored during the final push to the GitHub repository.

Submission of your GitHub repository link is mandatory. In order to track your task, you need to share the link of the repository in the document.

The step-by-step process involved in completing this task should be documented.